feat: make coordinate system more clear and consistent

+++ primitives.hpp
@@ -3,13 +3,13 @@
 class Vector2 {
 
 public:
-  float x, y;
+  double x, y;
 
-  Vector2(float _x, float _y);
+  Vector2(double _x, double _y);
 
   Vector2();
 
   Vector2 add(Vector2 v);
   Vector2 sub(Vector2 v);
-  Vector2 mul(float v);
+  Vector2 mul(double v);
 };
